104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,18 +9,19 @@
  */
 int main(void)
 {
-	unsigned int a = 1, b = 2, next;
+	unsigned long int a = 1, b = 2, next;
 	int count;
 
-	printf("%u, %u", a, b);
+	printf("%lu, %lu", a, b);
 
 	for (count = 2; count < 98; count++)
 	{
 		next = a + b;
-		printf(", %u", next);
+		printf(", %lu", next);
 		a = b;
 		b = next;
 	}
+
 	printf("\n");
 
 	return (0);
